requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
bootstrap4==0.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==40.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-bootstrap4==23.1
django-bootstrap5==23.1
django-crispy-forms==2.0
django-rosetta==0.9.8
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
polib==1.2.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
